Fixed FSUSER_IsSdmcDetected and FSUSER_IsSdmcWritable: the output is an u8 not u32.

+++ libctru/include/3ds/services/fs.h
@@ -144,8 +144,8 @@ Result FSUSER_DeleteDirectory(Handle *handle, FS_archive archive, FS_path dirLow
 Result FSUSER_RenameFile(Handle *handle, FS_archive srcArchive, FS_path srcFileLowPath, FS_archive destArchive, FS_path destFileLowPath);
 Result FSUSER_RenameDirectory(Handle *handle, FS_archive srcArchive, FS_path srcDirLowPath, FS_archive destArchive, FS_path destDirLowPath);
 Result FSUSER_GetSdmcArchiveResource(Handle *handle, u32 *sectorSize, u32 *clusterSize, u32 *numClusters, u32 *freeClusters);
-Result FSUSER_IsSdmcDetected(Handle *handle, u32 *detected);
-Result FSUSER_IsSdmcWritable(Handle *handle, u32 *writable);
+Result FSUSER_IsSdmcDetected(Handle *handle, u8 *detected);
+Result FSUSER_IsSdmcWritable(Handle *handle, u8 *writable);
 
 Result FSFILE_Close(Handle handle);
 Result FSFILE_Read(Handle handle, u32 *bytesRead, u64 offset, void *buffer, u32 size);
